+++ src/translations/translations.fr.js
@@ -131,7 +131,6 @@ export const translations = {
   'cc-addon-linked-apps.loading-error': `Une erreur est survenue pendant le chargement des applications liées.`,
   'cc-addon-linked-apps.no-linked-applications': `Aucune application liée pour l'instant.`,
   'cc-addon-linked-apps.title': `Applications liées`,
-  'cc-addon-linked-apps.zone': `zone:`,
   // cc-addon-mongodb-options
   'cc-addon-mongodb-options.description': `Sélectionnez les options que vous souhaitez pour votre add-on MongoDB.`,
   'cc-addon-mongodb-options.title': `Options pour l'add-on MongoDB`,